Converted to library, added to Homepage component
