+++ CancerGov/_src/Scripts/NCI/Modules/carousel/_carousel.scss
@@ -100,9 +100,9 @@
 }
 //.slick-slide {
 	/* background sprite for slides */
-	@import "../../../../StyleSheets/sprites/img-sprites";
-	@import "carousel-sprite-map";
-	@include img-sprites-same-size($carousel-sprites);
+	// @import "../../../../StyleSheets/sprites/img-sprites";
+	// @import "carousel-sprite-map";
+	// @include img-sprites-same-size($carousel-sprites);
 	/* End background sprite */
 //}
