+++ CancerGov/_src/Scripts/NCI/UX/PageSpecific/Home/HomePage.js
@@ -1,9 +1,14 @@
-define(function(require) {
-    require('./HomePage.scss');
-    require('Modules/carousel/carousel');
+import 'Modules/carousel/carousel';
+import espanolHome from 'Home/Enhancements/espanolHome';
+import CTDelighter from 'Common/Enhancements/clinicalTrialsDelighter';
+import patternInjector from 'Modules/patternInjector';
+import carouselSettings from './carouselSettings';
+import './HomePage.scss';
 
-    $(function() {
-        require('Home/Enhancements/espanolHome').init();
-        require('Common/Enhancements/clinicalTrialsDelighter').init();
-    });
-});
+
+const onDOMContentLoaded = () => {
+    espanolHome.init();
+    CTDelighter.init();
+}
+
+window.addEventListener('DOMContentLoaded', onDOMContentLoaded);
