+++ CancerGov/_src/Scripts/NCI/UX/PageSpecific/Home/carouselSettings.js
@@ -1,0 +1,137 @@
+const settings = {
+    '.blue-circles': {
+        str: '',
+        options: {
+            color: '',
+            generator: '',
+        }
+    },
+    '.blue-diamonds': {
+        str: '',
+        options: {
+            color: '',
+            generator: '',
+        }
+    },
+    '.blue-dots': {
+        str: '',
+        options: {
+            color: '',
+            generator: '',
+        }
+    },
+    '.blue-triangles': {
+        str: '',
+        options: {
+            color: '',
+            generator: '',
+        }
+    },
+    '.emerald-xes': {
+        str: '',
+        options: {
+            color: '',
+            generator: '',
+        }
+    },
+    '.fuchsia-waves': {
+        str: '',
+        options: {
+            color: '',
+            generator: '',
+        }
+    },
+    '.green-arrows': {
+        str: '',
+        options: {
+            color: '',
+            generator: '',
+        }
+    },
+    '.green-plaid': {
+        str: '',
+        options: {
+            color: '',
+            generator: '',
+        }
+    },
+    '.navy-squares': {
+        str: '',
+        options: {
+            color: '',
+            generator: '',
+        }
+    },
+    '.orange-plaid': {
+        str: '',
+        options: {
+            color: '',
+            generator: '',
+        }
+    },
+    '.orange-waves': {
+        str: '',
+        options: {
+            color: '',
+            generator: '',
+        }
+    },
+    '.pink-triangles': {
+        str: '',
+        options: {
+            color: '',
+            generator: '',
+        }
+    },
+    '.purple-circles': {
+        str: '',
+        options: {
+            color: '',
+            generator: '',
+        }
+    },
+    '.purple-triangles': {
+        str: '',
+        options: {
+            color: '',
+            generator: '',
+        }
+    },
+    '.red-squares': {
+        str: '',
+        options: {
+            color: '',
+            generator: '',
+        }
+    },
+    '.red-triangles': {
+        str: '',
+        options: {
+            color: '',
+            generator: '',
+        }
+    },
+    '.slate-grid': {
+        str: '',
+        options: {
+            color: '',
+            generator: '',
+        }
+    },
+    '.yellow-circles': {
+        str: '',
+        options: {
+            color: '',
+            generator: '',
+        }
+    },
+    '.yellow-hexagons': {
+        str: '',
+        options: {
+            color: '',
+            generator: '',
+        }
+    },
+}
+
+export default settings;
